Add max waiting connections to the listen function
The listen() function now takes an integer parameter max_waiting_connections that specifies the maximum number of waiting connections in the queue. This allows for better control over the connection backlog and improves the performance of the server. If listen() fails, the error message is printed to standard error and the program exits with a failure status.